Removed a note that was causing confusion.
The note said:

If you have two updates that happen within seconds of each other, then the flow may be triggered more than once with the latest versioned content.

This implies that if each individual action may not trigger a flow. 
I believe this was written to contrast with the polling behavior of the Dynamics 365 connector.