Update us-govt.md
Related to this PR for Power Apps: https://github.com/MicrosoftDocs/power-platform/pull/747

Article references a gateway whitelisting process and advises customers to open a support case to have it done. As far as I am aware (I'm in support) this is no longer necessary.

I am proposing a change to the gateway section to reflect this. Alternatively, you could argue that the entire gateway section could be deleted as there really is no difference in this functionality between commercial and US Gov customers.