Update get-started-with-copilot.md
Changed the availability table and a new explanation for "conversation flows (preview)" from the homepage.  

The experience is now shaped in 2 ways.  1 - if the pills are selected, particularly the third one "tell copilot what you want your flow to do..." it kicks off conversation first.

Otherwise, users can prompt like NL2FLOW and this will genarate a flow if there is enough info, or fall into the conversational copilot to ask for more info.