Actionable Approval mails in Outlook are not supported for Guest users in a tenant
As mentioned in this official document https://learn.microsoft.com/en-us/connectors/approvals/#known-issues-and-limitations
Actionable Approval mails in Outlook are not supported for Guest users in a tenant. Guest users will need to go to the Power Automate portal in order to act on an approval.
 
The sentence should be removed as it's causing confusion and misunderstanding.